way for dynamic entity addition
if the entity is added dynamically while its marker is visible to the camera the handler will not fire as isVisible is false. so you have to remove the camera from the marker (markerLost) then scan the marker again (markerFound) so that the events fire again and you are able to rotate and scale.
to circumvent it I copied isVisible from the parent marker data attributes. so if the user wants to add dynamic content and circumvent this events not firing on first scan he should add:
```
      AFRAME.registerComponent('markerhandler', {
        init: function () {

          this.el.sceneEl.addEventListener('markerLost', (e) => {
            e.target.dataset.isVisible = '0';
          });

          this.el.sceneEl.addEventListener('markerFound', (e) => {
            e.targetdataset.isVisible = '1';
          });
        }
      });
``` 
and on the marker:
```
  <a-marker markerhandler ...> ... </a-marker>
```